Issue 12: Extract assert_executes into its own assertions file
